split factories into _impl.h
